ApiKey: Implement last seen hook as a client hook instead
